requirements.txt
astroid==2.4.1
scipy
bs4
boto3
torch==1.5.1
torchvision==0.6.1
google-cloud-vision
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
distro==1.5.0
Flask==1.1.2
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.5
opencv-python==4.2.0.34
pandas==1.0.4
pdf2image==1.13.1
Pillow==7.1.2
pylint==2.5.2
PyPDF2==1.26.0
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
tabula-py==2.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
XlsxWriter==3.1.9
